+++ test/lib/active_record/associations_test.rb
@@ -8,4 +8,8 @@ describe AdjustableSchema::ActiveRecord::Associations do
 		_(Model1.new).must_respond_to :model2s
 		_(Model2.new).must_respond_to :model1s
 	end
+
+	describe 'inheritance' do
+		# TODO
+	end
 end